requirements.txt
numpy
opencv-python
Pillow
matplotlib
tk
pyserial
pyvisa
simple-pid 